test: restructure test suite into modular single-responsibility suites
Test architecture reorganized into domain-specific directories:

- Core/: CaptureActorTests, CaptureActorBatchTests, HeuristicRouterTests,
  RingBufferTests
- Helpers/: AuraConfigurationTests, AuraKitTests, TestHelpers
- Integration/: ConcurrencyTests, Phase2HardeningTests, StressTests
- Models/: SpatialEventCodableTests
- Persistence/: PersistenceHardeningTests, SwiftDataModelTests
- Security/: EncryptionServiceTests, KeyManagerTests, ZeroTrustTests
- Storage/: EncryptedMemoryStoreTests, EncryptedBatchAppendTests,
  MemoryStoreBatchTests, MemoryStoreTests, PaginatedStoreTests

21 test files across 7 modules — 168 tests, Thread Sanitizer verified

+++ Tests/AuraKitTests/Core/CaptureActorTests.swift
@@ -14,8 +14,8 @@ struct CaptureActorTests {
   // MARK: Helpers
 
   private func makeCaptureActor(
-    interactionWeight: Float = 1.0,
-    gazeWeight: Float = 0.3,
+    interactionWeight: Double = 1.0,
+    gazeWeight: Double = 0.3,
     bufferCapacity: Int = 64
   ) throws -> CaptureActor {
     let config = try AuraConfiguration(
